+++ src/main/kotlin/com/zepben/evolve/streaming/mutations/UpdateNetworkStateService.kt
@@ -10,15 +10,16 @@ package com.zepben.evolve.streaming.mutations
 
 import com.google.protobuf.Empty
 import com.zepben.evolve.conn.grpc.GrpcException
-import com.zepben.evolve.streaming.data.*
+import com.zepben.evolve.streaming.data.BatchFailure
+import com.zepben.evolve.streaming.data.CurrentStateEvent
+import com.zepben.evolve.streaming.data.SetCurrentStatesStatus
 import com.zepben.protobuf.connection.CheckConnectionRequest
 import com.zepben.protobuf.ns.SetCurrentStatesRequest
 import com.zepben.protobuf.ns.SetCurrentStatesResponse
 import com.zepben.protobuf.ns.UpdateNetworkStateServiceGrpc
 import io.grpc.stub.StreamObserver
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.sync.Mutex
-import java.util.concurrent.*
+import java.util.concurrent.CompletableFuture
+import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicInteger
 
 /**
@@ -64,16 +65,38 @@ class UpdateNetworkStateService(
      * back through the [responseObserver].
      */
     override fun setCurrentStates(responseObserver: StreamObserver<SetCurrentStatesResponse>): StreamObserver<SetCurrentStatesRequest> =
+
         object : StreamObserver<SetCurrentStatesRequest> {
-            private val outstandingProcesses = AtomicInteger()
-            private val onCompletedLock = Mutex()
+
+            /**
+             * An atomic int used for tracking when to send the `onCompleted` response.
+             *
+             * Since the callback used to handle requests can run on a separate thread, there are two possible places we may need to send the
+             * `onCompleted` response message:
+             * 1. In direct response to the `onCompleted` request if there are no outstanding callbacks.
+             * 2. When the last callback completes if a previous `onCompleted` request has been received.
+             *
+             * In order to prevent race conditions, this should be handled by a single variable. The race condition exists because the completable
+             * returned from `onSetCurrentStates` could be running on any thread.
+             *
+             * gRPC guarantees we will not receive the `onCompleted` request until after all `onNext` calls have been made. We guarantee that we
+             * increment this counter before we return from `onNext`, and decrement it when a callback is completed.
+             *
+             * By treating the `onCompleted` request to be the completion of an initial task (hence initialised to 1), it allows us to track when
+             * to send the `onCompleted` response, as both the completion of the last callback, and the handling of the `onCompleted` will decrement
+             * our counter.
+             *
+             * Given the only way of this number being zero is after all callbacks have been completed, and the `onCompleted` request being received,
+             * we can just check in the two places for a zero value as an indicator the `onCompleted` response needs to be sent.
+             */
+            private val completeWhenZero = AtomicInteger(1)
+
             override fun onNext(request: SetCurrentStatesRequest) {
                 try {
                     onSetCurrentStates(request.messageId, request.eventList.map { CurrentStateEvent.fromPb(it) })
                         .orTimeout(timeout, TimeUnit.SECONDS)
                         .also { completable ->
-                            if (outstandingProcesses.incrementAndGet() == 1)
-                                onCompletedLock.tryLock()
+                            completeWhenZero.incrementAndGet()
 
                             completable.whenComplete { result, e ->
                                 if (result != null)
@@ -81,8 +104,8 @@ class UpdateNetworkStateService(
                                 else if (e != null)
                                     handleError(request, "Error raised by the state updater", e)
 
-                                if (outstandingProcesses.decrementAndGet() == 0)
-                                    onCompletedLock.unlock()
+                                if (completeWhenZero.decrementAndGet() == 0)
+                                    responseObserver.onCompleted()
                             }
                         }
 
@@ -95,20 +118,21 @@ class UpdateNetworkStateService(
                 throw e
             }
 
-            override fun onCompleted() = runBlocking {
-                // Block response observer from completing until all outstanding processes are completed.
-                onCompletedLock.lock()
-                responseObserver.onCompleted()
+            override fun onCompleted() {
+                if (completeWhenZero.decrementAndGet() == 0)
+                    responseObserver.onCompleted()
             }
 
             private fun handleError(request: SetCurrentStatesRequest, message: String, e: Throwable) {
                 onProcessingError(GrpcException("$message for batch `${request.messageId}`: ${e.localizedMessage}", e))
                 responseObserver.onNext(BatchFailure(request.messageId, false, listOf()).toPb())
             }
+
         }
 
     override fun checkConnection(request: CheckConnectionRequest?, responseObserver: StreamObserver<Empty>) {
         responseObserver.onNext(Empty.getDefaultInstance())
         responseObserver.onCompleted()
     }
+
 }